Add files via upload
AmbuCall is a proposed digital ambulance hailing service created to respond to the emergency demand in ambulance hailing within urban cities. The app is going to link private and government owned emergency services on a location based database and connecting them to emergency request closest to them. The app intends to solve the challenges of delays and long waiting time that is caused by telecommunication operative redirecting calls and long driving distance.
AmbuCall will
Locate ambulance and medical center closest to user's location
2. Ambulance hailing and request
3. Notifies medical center/Physician of your emergency
4. Provides turn by turn directions, via the fastest route, to the closest available medical center
5. Provides a payment plan for emergency services